drivers: versal: ecc: Separate code using PLM firmware into its own file
In preparation for introducing a driver for Versal NET's PKI Accelerator
block.

Largely based on previous work by
Jeremie Corbier <jeremie.corbier@provenrun.com>.

Tested-by: Nathan Menhorn <nathan.menhorn@amd.com>
