+++ core/drivers/crypto/versal/sub.mk
@@ -3,4 +3,5 @@ incdirs-y += include
 srcs-y += ipi.c
 srcs-y += authenc.c
 srcs-y += ecc.c
+srcs-y += ecc_mbox.c
 srcs-y += rsa.c
